Fix incorrect assumptions about size_t.

Patches applied:

 * srivasta@debian.org--lenny/flex--debian--2.5.34--base-0
   tag of srivasta@debian.org--lenny/flex--upstream--2.5.34--patch-2

 * srivasta@debian.org--lenny/flex--debian--2.5.34--patch-1
   Do not assume unsigned int is the same as size_t


git-archimport-id: srivasta@debian.org--lenny/flex--devo--2.5.34--patch-1

+++ scan.c
@@ -181,11 +181,13 @@ extern FILE *yyin, *yyout;
 /* The following is because we cannot portably get our hands on size_t
  * (without autoconf's help, which isn't available because we want
  * flex-generated scanners to compile on their own).
+ * Given that the standard has decreed that size_t exists since 1989,
+ * I guess we can afford to depend on it. Manoj.
  */
 
 #ifndef YY_TYPEDEF_YY_SIZE_T
 #define YY_TYPEDEF_YY_SIZE_T
-typedef unsigned int yy_size_t;
+typedef size_t yy_size_t;
 #endif
 
 #ifndef YY_STRUCT_YY_BUFFER_STATE
